Fix data races (#45)
* Add -race flag to make test

* Fix data races

* Apply gofumpt to all Go files

* Fix some linter issues

* Fixes after code review

* Fix mutex usage

* Fix data race with `numLabelsWritten`

+++ proving/proving_test.go
@@ -50,7 +50,7 @@ func TestMain(m *testing.M) {
 func TestProver_GenerateProof(t *testing.T) {
 	r := require.New(t)
 
-	for numUnits := uint(cfg.MinNumUnits); numUnits < 6; numUnits++ {
+	for numUnits := cfg.MinNumUnits; numUnits < 6; numUnits++ {
 		newOpts := opts
 		newOpts.NumUnits = numUnits
